% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{adjust_pvalues}
\alias{adjust_pvalues}
\title{Adjust p-values within reporter}
\usage{
adjust_pvalues(
  table,
  pvalue = "specific_pvalue",
  output = "specific_padj_by_reporter",
  method = "BH"
)
}
\arguments{
\item{table}{Result table from \code{\link[=results]{results()}}.}

\item{pvalue}{P-value column.}

\item{output}{Name of adjusted p-value column.}

\item{method}{Passed to \code{\link[stats:p.adjust]{stats::p.adjust()}}.}
}
\description{
Adjust p-values within reporter
}
