% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{destress_presets}
\alias{destress_presets}
\title{List available DStressR presets}
\usage{
destress_presets()
}
\value{
A character vector of preset names accepted by \code{\link[=fit_destress]{fit_destress()}}.
}
\description{
List available DStressR presets
}
