% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{destress_workflows}
\alias{destress_workflows}
\title{List available DStressR workflows}
\usage{
destress_workflows()
}
\value{
A character vector of workflow names accepted by \code{\link[=fit_workflow]{fit_workflow()}}.
}
\description{
List available DStressR workflows
}
