% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_matrix_report}
\alias{model_matrix_report}
\title{Summarize model dimensions}
\usage{
model_matrix_report(fit)
}
\arguments{
\item{fit}{A \code{destress_fit}.}
}
\description{
Summarize model dimensions
}
