% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{read_campylobacter_expression}
\alias{read_campylobacter_expression}
\title{Read exported Campylobacter expression data}
\usage{
read_campylobacter_expression(expression_file, libmap_file)
}
\arguments{
\item{expression_file}{Path to \code{expression_values.tsv.gz}.}

\item{libmap_file}{Path to \code{LibMap.txt}.}
}
\value{
A data frame with \code{srn_code} and \code{ProductName} joined in.
}
\description{
Read exported Campylobacter expression data
}
