% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{results}
\alias{results}
\title{Extract model results}
\usage{
results(fit, perturbations = NULL, reporters = NULL)
}
\arguments{
\item{fit}{A \code{destress_fit} object.}

\item{perturbations}{Optional perturbation subset.}

\item{reporters}{Optional reporter subset.}
}
\value{
A data frame with total and reporter-specific effects.
}
\description{
Extract model results
}
