% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{results}
\alias{results}
\title{Extract model results}
\usage{
results(fit, compounds = NULL, promoters = NULL)
}
\arguments{
\item{fit}{A \code{destress_fit} object.}

\item{compounds}{Optional compound subset.}

\item{promoters}{Optional promoter subset.}
}
\value{
A data frame with total and promoter-specific effects.
}
\description{
Extract model results
}
